+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ## [0.7.0]
+### Added
+* editing object from relation modal is sending information to the live preview #26808
+
 ### Changed
 * open appliaction iframe on live preview button click #26808